+++ src/app/components/organizationview/organization-mails-view/organization-mails-view.component.ts
@@ -3,6 +3,8 @@ import {EmailTemplateModel} from "../../../models/EmailTemplateModel";
 import {MatSnackBar} from "@angular/material/snack-bar";
 import {InfoSnackbarComponent} from "./info-snackbar/info-snackbar.component";
 import {DataService} from "../../../services/DataService";
+import {MatDialog} from "@angular/material/dialog";
+import {DeletionConfirmationComponent} from "../../deletion-confirmation/deletion-confirmation.component";
 
 @Component({
   selector: 'app-organization-mails-view',
@@ -13,29 +15,12 @@ export class OrganizationMailsViewComponent {
   @Input() orgaID = '';
 
   availableEmailTemplates: EmailTemplateModel[] = [];
-  //availableEmailTemplates: EmailTemplateModel[] = [
-  //  {
-  //    id: '123',
-  //    name: 'Dankeschön',
-  //    subject: 'Liebe Teilnehmer Betreff',
-  //    text: 'Hallo zusammen, wir sind heute hier',
-  //    created: '12987',
-  //    lastModified: '56492h'
-  //  },
-  //  {
-  //    id: '789',
-  //    name: 'Dankeschön2',
-  //    subject: 'Liebe Menschen Betreff',
-  //    text: 'Hallo ihr doofians, moomjääj',
-  //    created: '12987',
-  //    lastModified: '56492h'
-  //  }
-  //];
+
   panelOpenState: boolean = false;
   editMode: boolean = false;
   editedEmailTemplate: string = '';
 
-  constructor(private dataService: DataService, private snackbar: MatSnackBar) {
+  constructor(private dataService: DataService, private snackbar: MatSnackBar, private dialog : MatDialog) {
   }
 
   ngOnInit(): void {
@@ -54,21 +39,22 @@ export class OrganizationMailsViewComponent {
   saveEmailTemplate(orgId: string, templateId: string, emailTemplate: EmailTemplateModel) {
     this.dataService.saveEmailTemplate(orgId, templateId, emailTemplate).subscribe(() => {
       this.ngOnInit();
-      this.snackbar.open('Rolle erfolgreich geändert', 'OK', {duration: 3000});
-    }, error => {
-      this.snackbar.open('Es ist ein Fehler aufgetreten', 'OK', {duration: 3000});
+      this.snackbar.open('Email Vorlage erfolgreich gespeichert', 'OK', {duration: 3000});
     })
     this.editMode = false;
     this.editedEmailTemplate = '';
   }
 
   deleteEmailTemplate(templateId: string) {
-    this.dataService.deleteMailTemplate(this.orgaID, templateId).subscribe(() => {
-      this.ngOnInit();
-      this.snackbar.open('Eintrag gelöscht', 'OK', {duration: 3000});
-    }, error => {
-      this.snackbar.open('Es ist ein Fehler aufgetreten', 'OK', {duration: 3000});
-    })
+    const dialogRef = this.dialog.open(DeletionConfirmationComponent,{});
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this.dataService.deleteMailTemplate(this.orgaID, templateId).subscribe(() => {
+          this.ngOnInit();
+          this.snackbar.open('Eintrag gelöscht', 'OK', {duration: 3000});
+        });
+      }
+    });
     console.log("delete Email Template " + templateId);
   }
